requirements.txt
PySide2
freecad-stubs
